Finish new schema structure
